test: moved testcontainers in base class
